+++ src/App.js
@@ -16,6 +16,13 @@ function App() {
 	const [subCount, setSubCount] = useState();
 	const banner = useRef();
 	const [viewStyle, setViewStyle] = useState(false);
+	const [shouldBlurAll, setShouldBlurAll] = useState(true);
+
+	useHotkeys("ctrl + shift + b", (e) => {
+		// prevent hiding, opening bookmarks bar.
+		e.preventDefault();
+		setShouldBlurAll((b) => !b);
+	});
 
 	// scroll to top
 	useHotkeys("g g", () => {
@@ -46,7 +53,8 @@ function App() {
 						previousSubreddit,
 						setSubCount,
 						banner,
-						viewStyle
+						viewStyle,
+						shouldBlurAll
 					}}
 				/>
 			</div>