+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 ## CHANGELOG
 
-1. changed sub selection input style, and press 'esc' to close it now.
+1. fixed no blur on over_18 single page load posts.
+2. ctrl + b to unblur post.
+3. ctrl + shift + b to unblur everything.